Break out submodules related into SubmoduleHelpers
