+++ ResourceManager/LocalizationHelpers.cs
@@ -112,7 +112,7 @@ namespace ResourceManager
         public static string ProcessSubmodulePatch(GitModule module, string fileName, Patch patch)
         {
             string text = patch?.Text;
-            var status = GitCommandHelpers.ParseSubmoduleStatus(text, module, fileName);
+            var status = SubmoduleHelpers.ParseSubmoduleStatus(text, module, fileName);
             if (status == null)
             {
                 return "";
